+++ xrpl/rpc/client.go
@@ -160,7 +160,10 @@ func (c *Client) SubmitTxBlob(txBlob string, failHard bool) (*requests.SubmitRes
 	if err != nil {
 		return nil, err
 	}
+	return c.submitTxBlob(txBlob, tx, failHard)
+}
 
+func (c *Client) submitTxBlob(txBlob string, tx map[string]any, failHard bool) (*requests.SubmitResponse, error) {
 	signingType, err := clientinternal.InspectSignedTransaction(tx, false)
 	if err != nil {
 		return nil, err
@@ -193,7 +196,7 @@ func (c *Client) SubmitTxBlobAndWait(txBlob string, failHard bool) (*requests.Tx
 		return nil, ErrMissingLastLedgerSequenceInTransaction
 	}
 
-	txResponse, err := c.SubmitTxBlob(txBlob, failHard)
+	txResponse, err := c.submitTxBlob(txBlob, tx, failHard)
 	if err != nil {
 		return nil, err
 	}
@@ -202,7 +205,7 @@ func (c *Client) SubmitTxBlobAndWait(txBlob string, failHard bool) (*requests.Tx
 		return nil, &ClientError{ErrorString: "transaction failed to submit with engine result: " + txResponse.EngineResult}
 	}
 
-	txHash, err := hash.SignTxBlob(txBlob)
+	txHash, err := hash.SignTx(tx)
 	if err != nil {
 		return nil, err
 	}
@@ -259,7 +262,7 @@ func (c *Client) SubmitMultisigned(txBlob string, failHard bool) (*requests.Subm
 		return nil, err
 	}
 	if signingType != clientinternal.MultiSignedTransaction {
-		return nil, ErrSignerDataIsEmpty
+		return nil, ErrTransactionNotMultisigned
 	}
 
 	return c.submitMultisignedRequest(&requests.SubmitMultisignedRequest{
@@ -276,14 +279,14 @@ func (c *Client) Autofill(tx *transaction.FlatTransaction) error {
 		return ErrNilTransaction
 	}
 	working := transaction.FlatTransaction(clientinternal.CloneTransaction(*tx))
-	if err := c.autofill(&working); err != nil {
+	if err := c.autofill(&working, 0); err != nil {
 		return err
 	}
 	clientinternal.ReplaceTransactionContents(*tx, working)
 	return nil
 }
 
-func (c *Client) autofill(tx *transaction.FlatTransaction) error {
+func (c *Client) autofill(tx *transaction.FlatTransaction, nSigners uint64) error {
 	if err := tx.RequireTransactionType(); err != nil {
 		return err
 	}
@@ -310,7 +313,7 @@ func (c *Client) autofill(tx *transaction.FlatTransaction) error {
 		}
 	}
 	if _, ok := (*tx)["Fee"]; !ok {
-		if err := c.calculateFeePerTransactionType(tx, 0); err != nil {
+		if err := c.calculateFeePerTransactionType(tx, nSigners); err != nil {
 			return err
 		}
 	}
@@ -345,10 +348,7 @@ func (c *Client) AutofillMultisigned(tx *transaction.FlatTransaction, nSigners u
 		return ErrNilTransaction
 	}
 	working := transaction.FlatTransaction(clientinternal.CloneTransaction(*tx))
-	if err := c.autofill(&working); err != nil {
-		return err
-	}
-	if err := c.calculateFeePerTransactionType(&working, nSigners); err != nil {
+	if err := c.autofill(&working, nSigners); err != nil {
 		return err
 	}
 	clientinternal.ReplaceTransactionContents(*tx, working)
@@ -403,93 +403,14 @@ func isFundWalletActNotFound(err error) bool {
 	return errors.As(err, &clientErr) && clientErr.ErrorString == actNotFound
 }
 
-type validatedInnerTx struct {
-	rawTx   map[string]any
-	account string
-}
-
 func (c *Client) autofillRawTransactions(tx *transaction.FlatTransaction) error {
-	identity, err := c.networkIdentity()
-	if err != nil {
-		return err
-	}
-	if err := clientinternal.ApplyNetworkIDPolicy(*tx, identity); err != nil {
-		return err
-	}
-
-	rawTxs, ok := (*tx)["RawTransactions"].([]map[string]any)
-	if !ok {
-		return ErrRawTransactionsFieldIsNotAnArray
-	}
-
-	inners := make([]validatedInnerTx, 0, len(rawTxs))
-	for _, rawTx := range rawTxs {
-		innerRawTx, ok := rawTx["RawTransaction"].(map[string]any)
-		if !ok {
-			return ErrRawTransactionFieldIsNotAnObject
+	return clientinternal.AutofillBatchRawTransactions(*tx, func(accountAddress string) (uint32, error) {
+		accountInfo, err := c.GetAccountInfo(&account.InfoRequest{
+			Account: types.Address(accountAddress),
+		})
+		if err != nil {
+			return 0, err
 		}
-
-		acc, ok := innerRawTx["Account"].(string)
-		if !ok {
-			return ErrAccountFieldIsNotAString
-		}
-
-		// Interface comparisons against a string literal are false for any
-		// non-string value (including nil), so no type assertion is needed.
-		if fee, exists := innerRawTx["Fee"]; exists && fee != "0" {
-			return types.ErrBatchInnerTransactionInvalid
-		}
-
-		if signingPubKey, exists := innerRawTx["SigningPubKey"]; exists && signingPubKey != "" {
-			return ErrSigningPubKeyFieldMustBeEmpty
-		}
-
-		if _, exists := innerRawTx["TxnSignature"]; exists {
-			return ErrTxnSignatureFieldMustBeEmpty
-		}
-		if _, exists := innerRawTx["Signers"]; exists {
-			return ErrSignersFieldMustBeEmpty
-		}
-
-		inners = append(inners, validatedInnerTx{rawTx: innerRawTx, account: acc})
-	}
-
-	accountSeq := make(map[string]uint32, len(inners))
-
-	for _, inner := range inners {
-		innerRawTx := inner.rawTx
-		if innerRawTx["Fee"] == nil {
-			innerRawTx["Fee"] = "0"
-		}
-
-		if innerRawTx["SigningPubKey"] == nil {
-			innerRawTx["SigningPubKey"] = ""
-		}
-
-		if innerRawTx["Sequence"] == nil && innerRawTx["TicketSequence"] == nil {
-			acc := inner.account
-
-			if accountSeq[acc] != 0 {
-				innerRawTx["Sequence"] = accountSeq[acc]
-				accountSeq[acc]++
-			} else {
-				accountInfo, err := c.GetAccountInfo(&account.InfoRequest{
-					Account: types.Address(acc),
-				})
-				if err != nil {
-					return err
-				}
-				var seq uint32
-				if innerRawTx["Account"] == (*tx)["Account"] {
-					seq = accountInfo.AccountData.Sequence + 1
-				} else {
-					seq = accountInfo.AccountData.Sequence
-				}
-				accountSeq[acc] = seq + 1
-				innerRawTx["Sequence"] = seq
-			}
-		}
-	}
-
-	return nil
+		return accountInfo.AccountData.Sequence, nil
+	})
 }